Add a decode API for embedders (ADR-0008)
Adapt the public interface to lambda1-grandprix's ADR-0003 so its λ-1
translator can observe result values directly, not via byte-stream
workarounds.

- `Program::eval_numeral() -> u64` reduces the program term to a Church
  numeral with no 256 cap (the `Acc` accumulator widens to 64-bit),
  replacing "emit N bytes and count" for decodeInt (their priority 1).
- `Program::eval_values(input, &DecodeOptions) -> Vec<u64>` reads the
  output `\f. f h t` list as raw Church numerals, not bytes, with a
  configurable/disable-able EOF sentinel and element/step bounds — covers
  flat strings and integer arrays (their priorities 2-flat, 3).
- `Program::from_term(Term)` runs an already-built AST, and `Term`/`Comb`
  are re-exported at the crate root, skipping the render/re-parse trip
  (their priority 4).
- `church_numeral(n) -> Term` builds a host integer as a native numeral
  (their priority 6).

Priorities 5 (bounded eval) and 7 (safe failure) were already met by
ADR-0007. Nested Scott-list decode (decodeJson) is deferred — it needs
term-handle traversal. Byte-stream I/O is unchanged (differential fuzz:
0 mismatches); all 39 tests pass.

+++ src/vm/reduce.rs
@@ -82,7 +82,7 @@ impl Vm {
                     if matches!(self.resolve(f), Cell::Comb(Comb::Inc))
                         && let Cell::Acc(k) = self.resolve(x)
                     {
-                        let acc = self.alloc(Cell::Acc(k + nv));
+                        let acc = self.alloc(Cell::Acc(k + nv as u64));
                         self.heap.set(redex, Cell::Ind(acc));
                     } else if nv == 0 {
                         self.heap.set(redex, Cell::Ind(x)); // 0 f x = x
